sendero_unillanos/requirements.txt
Django==2.2
django-singleton-model==0.2.3
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
